rename file to be more verbose, replace q with bluebird, get rid of defer anti pattern
